Lot of suff updated, some things left undone in gnu enviroment ==============================================================
System Information:
Engine Version: Molecule Ver 1.6.4.20220108
Library Version:  Ver 2.3.2.20220108

Hosts:
Interactive Console Ver 1.0.5.20220108
Pipe Child Ver 1.1.2.20220108
Headless Host Ver 1.0.0.20220108

Storages:
File System Storage Ver 1.1.2.20220108

Firmware:
Windows Ver 1.2.0.20220107
Linux Ver 1.0.4.20220108

Tested Compilers:
MSVC 14.29.30138
GCC 9.3.0

Tested OSes:
Linux Ver:5.11.0-44-generic [x86_64]
Windows Workstation 10.0.19044 [x64]

+++ common/rx_value_conv.c
@@ -4,27 +4,27 @@
 *
 *  common\rx_value_conv.c
 *
-*  Copyright (c) 2020-2021 ENSACO Solutions doo
+*  Copyright (c) 2020-2022 ENSACO Solutions doo
 *  Copyright (c) 2018-2019 Dusan Ciric
 *
-*
+*  
 *  This file is part of {rx-platform}
 *
-*
+*  
 *  {rx-platform} is free software: you can redistribute it and/or modify
 *  it under the terms of the GNU General Public License as published by
 *  the Free Software Foundation, either version 3 of the License, or
 *  (at your option) any later version.
-*
+*  
 *  {rx-platform} is distributed in the hope that it will be useful,
 *  but WITHOUT ANY WARRANTY; without even the implied warranty of
 *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *  GNU General Public License for more details.
-*
-*  You should have received a copy of the GNU General Public License
+*  
+*  You should have received a copy of the GNU General Public License  
 *  along with {rx-platform}. It is also available in any {rx-platform} console
 *  via <license> command. If not, see <http://www.gnu.org/licenses/>.
-*
+*  
 ****************************************************************************/
 
 
@@ -345,7 +345,7 @@ int int32_to_str(int32_t val, string_value_struct* str)
 int int64_to_str(int64_t val, string_value_struct* str)
 {
 	char buff[0x20];
-	sprintf(buff, "%d", (int)val);
+	sprintf(buff, "%lld", val);
 	return rx_init_string_value_struct(str, buff, -1);
 }
 
@@ -364,13 +364,13 @@ int uint16_to_str(uint16_t val, string_value_struct* str)
 int uint32_to_str(uint32_t val, string_value_struct* str)
 {
 	char buff[0x10];
-	sprintf(buff, "%u", (int)val);
+	sprintf(buff, "%u", val);
 	return rx_init_string_value_struct(str, buff, -1);
 }
 int uint64_to_str(uint64_t val, string_value_struct* str)
 {
 	char buff[0x20];
-	sprintf(buff, "%u", (int)val);
+	sprintf(buff, "%llu", val);
 	return rx_init_string_value_struct(str, buff, -1);
 }
